bhyve/tpm_ppi_qemu: Cast address of the TPP3 field to uint32_t
Fix a format specifier issue when building 'tpm_ppi_qemu.c' with
'dsdt_line' marked as 'printflike'.

Reviewed by:	markj, imp
Differential Revision:	https://reviews.freebsd.org/D51535

+++ usr.sbin/bhyve/tpm_ppi_qemu.c
@@ -207,7 +207,7 @@ tpm_ppi_write_dsdt_regions(void *sc __unused)
 	 * Used for TCG Platform Reset Attack Mitigation
 	 */
 	dsdt_line("OperationRegion(TPP3, SystemMemory, 0x%8x, 1)",
-	    TPM_PPI_ADDRESS + sizeof(struct tpm_ppi_qemu));
+	    TPM_PPI_ADDRESS + (uint32_t)sizeof(struct tpm_ppi_qemu));
 	dsdt_line("Field(TPP3, ByteAcc, NoLock, Preserve)");
 	dsdt_line("{");
 	dsdt_line("  MOVV, 8,");
